requirements.txt
tensorflow==2.1.0
pickle
numpy
Pillow